Add CammandType.java Main.java: Add error handling
System would crash whenever an input was unknown or format was incorrect

Implemented a new function called inputCheck to check for any descriptions present as well as correct formatting.
CommandType class was created to abstract the different formats of commands.

Using objects made the code cleaner and reduced repetitive code.